+++ TeamCode/src/main/java/org/firstinspires/ftc/team7316/util/input/GamepadWrapper.java
@@ -13,8 +13,8 @@ public class GamepadWrapper implements Loopable {
 
     private Gamepad gamepad;
 
-    private float multLeft = 1f;
-    private float multRight = 1f;
+    private double multLeft = 1;
+    private double multRight = 1;
 
     public JoystickWrapper left_stick, right_stick;
     public AxisWrapper left_axis_y, right_axis_y;
@@ -84,7 +84,7 @@ public class GamepadWrapper implements Loopable {
         throw new IllegalArgumentException();
     }
 
-    public float axisValue(GamepadAxis axisIndex) {
+    public double axisValue(GamepadAxis axisIndex) {
         switch (axisIndex) {
             case L_STICK_X: return gamepad.left_stick_x;
             case L_STICK_Y: return deadzone(-gamepad.left_stick_y, multLeft);
@@ -96,17 +96,17 @@ public class GamepadWrapper implements Loopable {
         throw new IllegalArgumentException();
     }
 
-    private float squared(float value) {
-        float result = (float) Math.abs(Math.pow(value, 2));
+    private double squared(double value) {
+        double result = (double) Math.abs(Math.pow(value, 2));
         return value > 0 ? result : -result;
     }
 
-    private float deadzone(float value, float mult) {
+    private double deadzone(double value, double mult) {
         if (Math.abs(value) < Constants.JOYSTICK_DRIVE_DEADZONE) {
             return 0;
         } else {
-            float slope = (1 - Constants.DRIVER_MOTOR_DEADZONE) / (1 - Constants.JOYSTICK_DRIVE_DEADZONE) * mult;
-            float preval = slope * (Math.abs(value) - Constants.JOYSTICK_DRIVE_DEADZONE) + Constants.DRIVER_MOTOR_DEADZONE;
+            double slope = (1 - Constants.DRIVER_MOTOR_DEADZONE) / (1 - Constants.JOYSTICK_DRIVE_DEADZONE) * mult;
+            double preval = slope * (Math.abs(value) - Constants.JOYSTICK_DRIVE_DEADZONE) + Constants.DRIVER_MOTOR_DEADZONE;
 
             if (value > 0) {
                 return preval;
@@ -126,13 +126,13 @@ public class GamepadWrapper implements Loopable {
     @Override
     public void loop() {
         if (this.right_bumper.state()) {
-            multRight = 0.1f;
+            multRight = 0.1;
         } else {
             multRight = 1;
         }
 
         if (this.left_bumper.state()) {
-            multLeft = 0.1f;
+            multLeft = 0.1;
         } else {
             multLeft = 1;
         }
